+++ example/integration_test/all_integration_tests.dart
@@ -3,9 +3,11 @@ import 'package:flutter_bdd_suite/integration_test_helper.dart';
 import 'test_config.dart';
 
 import 'generated/auth/login.dart' as login;
+import 'generated/dashboard/employee_directory.dart' as employee_directory;
 
 void main() {
   IntegrationTestHelper(config: config);
 
   login.main();
+  employee_directory.main();
 }